Improve short-circuit mechanism.
Previously I had a boolean to keep track of any swaps, if none were
performed on a given pass the sort was considered done.

Here I change the logic to not just keep track of if a swap has occurred
but where it occurs. If a bunch of values are already in order the sort
can shorten the number of passes to only work up until the position of
the last swap.